Updated persistent memory code

+++ src/led.h
@@ -1,6 +1,8 @@
 /*
  * led.h
  *
+ * Corresponding header file for led.c
+ *
  *  Created on: 18-Nov-2019
  *      Author: Devansh
  */
@@ -8,8 +10,10 @@
 #ifndef SRC_LED_H_
 #define SRC_LED_H_
 
+#include <stdint.h>
+
 void test_led(void);
 void timer_pwm();
-void set_duty_cycle(int duty_cycle);
+void set_duty_cycle(uint8_t duty_cycle);
 
 #endif /* SRC_LED_H_ */
